[sc-11053] [refactor] Moved the EventTableViewController to its own file. The app now alerts the user at startup with an explanation of what it does.
